requirements.txt
requests
fastapi>=0.110
uvicorn[standard]>=0.23
sqlalchemy>=2.0
alembic>=1.12
psycopg2-binary>=2.9
slowapi>=0.1.9
python-dotenv>=1.0
PyJWT>=2.8
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.9
email-validator>=2.0